Add files via upload
Please be aware that this architecture is designed to function seamlessly when the stored procedure is initiated by the Lambda or Glue code itself, as is the case in our context. However, if the stored procedures are manually triggered by individuals, a potential issue arises. In scenarios where multiple data sources are being merged into a single source, Snowflake will only update a single file log if multiple individuals activate the stored procedure concurrently.